+++ vulture_allowlist.py
@@ -1,3 +1,5 @@
+from agent_plugin_builder.agent_plugin_builder import CustomFormatter
 from agent_plugin_builder.build_options import PlatformDependencyPackagingMethod
 
+CustomFormatter._get_help_string
 PlatformDependencyPackagingMethod.AUTODETECT