+++ CHANGELOG.md
@@ -12,7 +12,10 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 - `--verify/--no-verify` CLI option. #8
 
 ### Removed
-- plugin's build.yml configuration scripts. #8
+- Plugin's build.yml configuration scripts. #8
+
+### Changed
+- Default value for `--platform-dependencies` CLI option to "autodetect". #9
 
 ## [0.4.0] - 2024-07-18
 ### Fixed