+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
-## [Unreleased]
+## [0.5.0] - 2024-07-22
 ### Added
 - `--platform-dependencies` CLI option. #8
 - `--verify/--no-verify` CLI option. #8